modularize more, debug mode, optuna compatible network.
